+++ .github/workflows/ci-python.yml
@@ -18,7 +18,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install black
-          pip install -r python-backend/requirements.txt
+          pip install -r python_backend/requirements.txt
       - name: Check formatting
         run: black --check .
       - name: Run tests
