+++ Dockerfile
@@ -13,11 +13,11 @@ FROM python:3.11-slim as backend
 WORKDIR /app
 
 # install python dependencies
-COPY python-backend/requirements.txt ./
+COPY python_backend/requirements.txt ./
 RUN pip install --no-cache-dir -r requirements.txt
 
 # copy backend source
-COPY python-backend ./python-backend
+COPY python_backend ./python_backend
 
 # copy built frontend from previous stage
 COPY --from=frontend /app/.next ./frontend/.next
@@ -25,4 +25,4 @@ COPY --from=frontend /app/public ./frontend/public
 COPY --from=frontend /app/next.config.js ./frontend/next.config.js
 
 EXPOSE 8000
-CMD ["uvicorn", "main:app", "--app-dir", "python-backend", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn", "main:app", "--app-dir", "python_backend", "--host", "0.0.0.0", "--port", "8000"]
