+++ python_backend/routes/roof.py
